update tagId as int to long

+++ src/main/java/entity/Tag.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Tag {
-    private int tagId;
+    private long tagId;
     private String title;
 }
